+++ x/incentive/keeper/msg_server_delegator_test.go
@@ -47,7 +47,7 @@ func (suite *HandlerTestSuite) TestPayoutDelegatorClaimMultiDenom() {
 
 	// Claim denoms
 	err := suite.DeliverIncentiveMsg(&msg)
-	suite.NoError(err)
+	suite.Require().NoError(err)
 
 	// Check rewards were paid out
 	expectedRewardsHard := c("hard", int64(0.2*float64(2*7*1e6)))
@@ -97,7 +97,7 @@ func (suite *HandlerTestSuite) TestPayoutDelegatorClaimSingleDenom() {
 
 	// Claim rewards
 	err := suite.DeliverIncentiveMsg(&msg)
-	suite.NoError(err)
+	suite.Require().NoError(err)
 
 	// Check rewards were paid out
 	expectedRewards := c("swap", 2*7*1e6)