+++ x/incentive/keeper/rewards_borrow_test.go
@@ -86,7 +86,7 @@ func (suite *BorrowIntegrationTests) TestSingleUserAccumulatesRewardsAfterSyncin
 	})
 
 	// User claims all their rewards
-	suite.NoError(suite.DeliverIncentiveMsg(&msg))
+	suite.Require().NoError(suite.DeliverIncentiveMsg(&msg))
 
 	// The users has always had 100% of borrows, so they should receive all rewards for the previous two blocks.
 	// Total rewards for each block is block duration * rewards per second